Refatora service e controller do localizador
