+++ src/routes/locator.routes.ts
@@ -5,6 +5,6 @@ const locatorRouter = Router();
 const locatorController = new LocatorController();
 
 locatorRouter.get('/', locatorController.index);
-locatorRouter.get('/:code', locatorController.getLocator);
+locatorRouter.get('/:code', locatorController.show);
 
 export default locatorRouter;
